Massive rework of postgres extension handling.
Rather than being inside the summon controller, this is now part of postgres database. This is mostly required because you need admin creds to run the extension install in most cases, but we no longer expose the admin creds to the summon controller (which seems correct). This also makes it easier to install extensions for microservices some day.